requirements.txt
# Core data processing
akshare>=1.14.0
pandas>=2.0.0
pyarrow>=14.0.0

# Database and storage
duckdb>=0.10.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
